[MIM-37] PlanTab에서 계획 추가 구현, HomeTab에서 추가된 계획 보이도록 구현
- PlanTab에서 계획 추가 구현, HomeTab에서 추가된 계획 보이도록 구현
- 각 useCase생성
- planRepository구현
- ViewModel 구현하여 UI에서 호출하기
- observe하여 UI 변경
- PlanState enum class 생성하여 값 사용하기

+++ data/src/main/java/com/moon/data/room/dao/PlanDao.kt
@@ -29,7 +29,7 @@ interface PlanDao {
         update plan_table set plan_date = :newDate, plan_state = "LATER"
         where plan_id = :planId
     """)
-    suspend fun setPlanDelayOneDay(planId: Int, newDate: Date)
+    suspend fun setPlanDelayOneDay(planId: Long, newDate: Date)
 
     // 5. 해당 날짜의 계획 조회하기
     @Query(
